### 2020-08-14 #2566 @diehlpk
Hey @JaroslavHron @tupi @wkearn @dlagrava @adam-m-jcbs @mzamoraz @ctdegroot @mancellin @zbeekman would you be interested to review this paper?
