### 2020-08-14 #2566 @diehlpk
> ```
> Reference check summary:
>
> OK DOIs
>
> - 10.11588/ans.2015.100.20553 is OK
> - 10.1145/2566630 is OK
> - 10.1016/j.jfluidstructs.2003.12.004 is OK
> - 10.1017/jfm.2017.910 is OK
> - 10.1016/j.jfluidstructs.2015.12.009 is OK
> - 10.1093/jxb/erz288 is OK
> - 10.1146/annurev-fluid-122414-034606 is OK
> - 10.1098/rstb.1998.0262 is OK
> - 10.5281/zenodo.1287832 is OK
> - 10.1016/j.jfluidstructs.2015.12.009 is OK
> - 10.1016/j.jfluidstructs.2015.11.007 is OK
> - 10.1017/S0022112009993673 is OK
>
> MISSING DOIs
>
> - https://doi.org/10.5860/choice.48-5112 may be missing for title: Elasticity and Geometry: From hair curls to the nonlinear response of shells
> - https://doi.org/10.1146/annurev.fl.09.010177.002011 may be missing for title: Fluid mechanics of propulsion by cilia and flagella
> - https://doi.org/10.1073/pnas.1409118111 may be missing for title: Coiling of elastic rods on rigid substrates
>
> INVALID DOIs
>
> - None
> ```

 @mou3adb Could you please check the missing DOIs and add them if they are missing?
