### 2020-08-14 #2566 @diehlpk
@mou3adb Could you please add Montreal as city to both affiliations? Even it is mentioned in the school's name, we need the city as well.
