### 2020-08-14 10251f1
added doi and city
