+++ paper/paper.bib
@@ -2,7 +2,6 @@
 	address = {Oxford ; New York},
 	title = {Elasticity and {Geometry}: {From} hair curls to the nonlinear response of shells},
 	isbn = {978-0-19-850625-6},
-	shorttitle = {Elasticity and {Geometry}},
 	language = {English},
 	publisher = {Oxford University Press},
 	author = {Audoly, Basile and Pomeau, Yves},
@@ -118,7 +117,6 @@
 @book{paidoussis_fluid-structure_2010,
 	title = {Fluid-{Structure} {Interactions}: {Cross}-{Flow}-{Induced} {Instabilities}},
 	isbn = {978-1-139-49190-7},
-	url = {https://books.google.ca/books?id=9BZ_BiOGD0sC},
 	publisher = {Cambridge University Press},
 	author = {Païdoussis, M.P. and Price, S.J. and {de Langre}, Emmanuel},
 	year = {2010}
@@ -213,6 +211,7 @@
 	volume = {9},
 	number = {1},
 	journal = {Annual Review of Fluid Mechanics},
+	doi = {10.1146/annurev.fl.09.010177.002011},
 	author = {Brennen, Christopher and Winet, Howard},
 	year = {1977},
 	note = {Publisher: Annual Reviews 4139 El Camino Way, PO Box 10139, Palo Alto, CA 94303-0139, USA},
@@ -239,6 +238,7 @@
   author={Jawed, Mohammad K and Da, Fang and Joo, Jungseock and Grinspun, Eitan and Reis, Pedro M},
   journal={Proceedings of the National Academy of Sciences},
   pages={201409118},
+  doi = {10.1073/pnas.1409118111},
   year={2014},
   publisher={National Acad Sciences}
 }
