### 2020-08-14 #2566 @mou3adb
Hi @diehlpk. I added the city, and fixed the DOIs, except for the book Elasticity and Geometry since it already has an ISBN.
