### 2020-08-16 #2577 @whedon
[REVIEW]: RodiCS: a finite element solver of Kirchhoff rods under fluid flow and more
**Submitting author:** @mou3adb (<a href="http://orcid.org/0000-0002-4908-4589">Mouad Boudina</a>)
**Repository:** <a href="https://github.com/lm2-poly/RodiCS" target ="_blank">https://github.com/lm2-poly/RodiCS</a>
**Version:** v1.0.0
