+++ paper/paper.bib
@@ -159,10 +159,9 @@
 	year = {2020},
 }
 
-@article{boudina_jfm_2020,
+@article{boudina_jfm,
 	title = {Vortex-induced vibrations: a soft coral feeding strategy?},
-	language = {en},
-	journal = {Submitted to the Journal of Fluid Mechanics},
+	journal = {Journal of Fluid Mechanics (submitted)},
 	author = {Boudina, Mouad and Gosselin, Frédérick P. and Étienne, Stéphane},
 	month = dec,
 	year = {2020},
@@ -241,4 +240,61 @@
   doi = {10.1073/pnas.1409118111},
   year={2014},
   publisher={National Acad Sciences}
+}
+
+@article{violette_computation_2007,
+	title = {Computation of vortex-induced vibrations of long structures using a wake oscillator model: {Comparison} with {DNS} and experiments},
+	volume = {85},
+	issn = {0045-7949},
+	shorttitle = {Computation of vortex-induced vibrations of long structures using a wake oscillator model},
+	url = {http://www.sciencedirect.com/science/article/pii/S004579490600229X},
+	doi = {10.1016/j.compstruc.2006.08.005},
+	number = {11},
+	journal = {Computers \& Structures},
+	author = {Violette, R. and de Langre, E. and Szydlowski, J.},
+	month = jun,
+	year = {2007},
+	pages = {1134--1141},
+}
+
+@article{leclercq_vortex-induced_2018,
+	title = {Vortex-induced vibrations of cylinders bent by the flow},
+	volume = {80},
+	issn = {0889-9746},
+	url = {http://www.sciencedirect.com/science/article/pii/S0889974617307764},
+	doi = {10.1016/j.jfluidstructs.2018.03.008},
+	journal = {Journal of Fluids and Structures},
+	author = {Leclercq, Tristan and de Langre, Emmanuel},
+	month = jul,
+	year = {2018},
+	pages = {77--93},
+}
+
+@article{lei_wave_2019,
+	title = {Wave damping by flexible vegetation: {Connecting} individual blade dynamics to the meadow scale},
+	volume = {147},
+	issn = {03783839},
+	shorttitle = {Wave damping by flexible vegetation},
+	url = {https://linkinghub.elsevier.com/retrieve/pii/S0378383918300905},
+	doi = {10.1016/j.coastaleng.2019.01.008},
+	urldate = {2019-04-25},
+	journal = {Coastal Engineering},
+	author = {Lei, Jiarui and Nepf, Heidi},
+	month = may,
+	year = {2019},
+	pages = {138--148},
+}
+
+@article{leclercq_drag_2016,
+	title = {Drag reduction by elastic reconfiguration of non-uniform beams in non-uniform flows},
+	volume = {60},
+	issn = {08899746},
+	url = {http://linkinghub.elsevier.com/retrieve/pii/S088997461500242X},
+	doi = {10.1016/j.jfluidstructs.2015.10.007},
+	urldate = {2015-12-13},
+	journal = {Journal of Fluids and Structures},
+	author = {Leclercq, Tristan and de Langre, Emmanuel},
+	month = jan,
+	year = {2016},
+	pages = {114--129},
 }
