+++ paper/paper.md
@@ -27,25 +27,25 @@ bibliography: paper.bib
 
 # Summary
 
-Elastic rods are long and thin structures, and are the object of many dynamical systems in nature, from bending and swaying plants [@hassani_large_2016; @luhar_wave-induced_2016; @gosselin_mechanics_2019] to swimming flagellated and ciliated bacteria [@brennen_fluid_1977; @lauga_bacterial_2016] and propulsing carangiform fish [@cheng_continuous_1998]. In industry as well, elastic rods serve as models to study drilling risers in offshore sites and tall chimney stacks in constructions that may vibrate in presence of water or air flows [@paidoussis_fluid-structure_2010]. With the prevalence of rods in mechanical problems, it is important to have a simple to use efficient numerical tool reproducing faithfully the observed dynamics. RodiCS is a Python code that fulfills this task, and simulates the motion of elastic rods under specified forces in general, and their interaction with fluid flows more specifically.
-
+Elastic rods are long and thin structures, and are the object of many dynamical systems in nature, from bending and swaying plants [@gosselin_mechanics_2019] to swimming flagellated and ciliated bacteria [@brennen_fluid_1977; @lauga_bacterial_2016] and propulsing carangiform fish [@cheng_continuous_1998]. In industry as well, elastic rods serve as models to study drilling risers in offshore sites and tall chimney stacks in constructions that may vibrate in presence of water or air flows [@paidoussis_fluid-structure_2010]. With the prevalence of rods in mechanical problems, it is important to have a simple to use efficient numerical tool reproducing faithfully the observed dynamics. RodiCS is a Python code that fulfills this task, and simulates the motion of elastic rods under defined forces in general, and their interaction with fluid flows more specifically. While numerous researchers [@violette_computation_2007; @lei_wave_2019; @gosselin_drag_2010; @hassani_large_2016; @leclercq_reconfiguration_2018; @leclercq_vortex-induced_2018; @leclercq_drag_2016; @luhar_wave-induced_2016] spend valuable time in writing *ad hoc* codes with limited verification and validation, RodiCS comes as a versatile and practical open source library that scientific projects can benefit from and build upon.
+ 
 # Description
 
-With the FEniCS platform [@alnaes_fenics_2015], RodiCS solves the Kirchhoff equations of elastic rods [@landau_theory_1986; @audoly_elasticity_2010] using finite elements. The first step in each simulation is to import the mesh of the rod from an `.xml` file. The mesh consists of one-dimensional first order Lagrange elements embedded in a three-dimensional space. Then, external forces applied on the rod need to be listed. In FEniCS, the Unified Form Language (UFL) [@alnaes_unified_2014] ensures a convenient syntax to write mathematical expressions. Algebraic, vectorial, and differential operators, as well as tensor algebra and index notation, are all included in UFL, so users can enter force expressions exactly as they figure in equations without any difficulty. After applying boundary conditions, Kirchhoff equations are solved with the Newton method. In the case of transient simulations, these equations are discretised in time using the backward Euler scheme. \autoref{fig:examples}(a) shows the static reconfiguration of a 2D beam under a semi-empirical drag formulation for several load magnitudes, whereas \autoref{fig:examples}(b) records the dynamic deformation of a similar beam under an oscillatory flow over a period. Numerical results match well experiments on thin plates, which are also governed by Kirchhoff equations, conducted by @gosselin_drag_2010 and @leclercq_reconfiguration_2018. A more detailed description of Kirchhoff equations, solving method, and code validation can be found in @boudina_corals_2020.
[…]
