### 2020-09-01 ae466fa
paragraph DER and pdf of thesis
