+++ paper/paper.md
@@ -37,9 +37,9 @@ With the FEniCS platform [@alnaes_fenics_2015], RodiCS solves the Kirchhoff equa
 
 # Statement of Need
 
-The traditional method to simulate the motion of elastic rods in FEniCS is to solve the elasticity equation in three dimensions. This method takes a mesh with a large number of volumetric elements, hence is costly, time-consuming, and particularly inappropriate for slender structures. A more efficient option is to consider a rod theory that describes deformation based only on the curvilinear coordinate of the rod, which transposes numerically into using one-dimensional elements in three-dimensional space. Such a code was proposed by @bleyer_numerical_2018 solving Timoshenko beams. However, it is restricted to static deformations, and is valid only for small displacements due to the linearity of equations. In RodiCS, we simulate both static and dynamic deformations of Kirchhoff rods, which are more representative of thin rods, and we account for large displacements owing to its nonlinear character. RodiCS enters then as an important contribution to open source finite element solvers, as free software seldom use one-dimensional Kirchhoff rod elements embedded in a three-dimensional space.
+The traditional method to simulate the motion of elastic rods in FEniCS is to solve the elasticity equation in three dimensions. This method takes a mesh with a large number of volumetric elements, hence is costly, time-consuming, and particularly inappropriate for slender structures. A more efficient option is to consider a rod theory that describes deformation based only on the curvilinear coordinate of the rod, which transposes numerically into using one-dimensional elements in three-dimensional space. Such a code was proposed by @bleyer_numerical_2018 solving Timoshenko beams. However, it is restricted to static deformations, and is valid only for small displacements due to the linearity of equations. In RodiCS, we simulate both static and dynamic deformations of Kirchhoff rods, which are more representative of thin rods, and we account for large displacements owing to its nonlinear character. Therefore, RodiCS fills a need in open source finite element codes for a way to solve the one-dimensional Kirchhoff rod equation embedded in a three-dimensional space.
 
-Another elastic rod solver worth to mention is the Discrete Elastic Rods (DER) code of @jd2014. Because the DER code is written in C++ and has an intricate file structure, only programmer users are able to modify the source code, excluding thus users new in coding or having basic knowledge in programming. Moreover, it is not cross-platform and has several dependencies, limiting again access to a number of users. Last but not least, @jd2014 acknowledge themselves that they cannot maintain the code or provide support anymore. These drawbacks were incentives for the authors of the present paper to think of writing their own code based on the FEniCS project, which is cross-platform, actively maintained, continuously developed, and accessible for users of all programming levels.
+Another elastic rod solver worth to mention is the Discrete Elastic Rods (DER) code of @jd2014. Because the DER code is written in C++ and has an intricate file structure, only programmer users are able to modify the source code, excluding thus users new in coding or having basic knowledge in programming. Moreover, it is not cross-platform and has several dependencies, limiting again access to a number of users. Last but not least, @jd2014 acknowledge themselves that they cannot maintain the code or provide support anymore. These drawbacks were incentives for the authors of the present paper to think of writing their own code based on the FEniCS project, an open source finite element solver with an active community, continuously developed, and cross-platform; all significant advantages. Also as an essential modular feature, FEniCS imposes a standard in the mesh definition, making possible either the mesh generation with external meshing software, or the combination of elements in RodiCS with other types of finite elements. Finally, written in Python, RodiCS is inclusive and accessible for users of all programming levels.
 
[…]
