+++ paper/paper.bib
@@ -162,7 +162,7 @@
 @article{boudina_jfm,
 	title = {Vortex-induced vibrations: a soft coral feeding strategy?},
 	journal = {Journal of Fluid Mechanics (submitted)},
-	author = {Boudina, Mouad and Gosselin, Frédérick P. and Étienne, Stéphane},
+	author = {Boudina, Mouad and Gosselin, Frédérick P. and Etienne, Stéphane},
 	month = dec,
 	year = {2020},
 }
@@ -226,7 +226,7 @@
 	language = {en},
 	urldate = {2015-11-05},
 	journal = {Journal of Fluid Mechanics},
-	author = {Gosselin, F. and {de Langre}, Emmanuel and Machado-Almeida, Bruno A.},
+	author = {Gosselin, Frédérick and {de Langre}, Emmanuel and Machado-Almeida, Bruno A.},
 	month = may,
 	year = {2010},
 	pages = {319},
@@ -251,7 +251,7 @@
 	doi = {10.1016/j.compstruc.2006.08.005},
 	number = {11},
 	journal = {Computers \& Structures},
-	author = {Violette, R. and de Langre, E. and Szydlowski, J.},
+	author = {Violette, R. and {de Langre}, E. and Szydlowski, J.},
 	month = jun,
 	year = {2007},
 	pages = {1134--1141},
@@ -264,7 +264,7 @@
 	url = {http://www.sciencedirect.com/science/article/pii/S0889974617307764},
 	doi = {10.1016/j.jfluidstructs.2018.03.008},
 	journal = {Journal of Fluids and Structures},
-	author = {Leclercq, Tristan and de Langre, Emmanuel},
+	author = {Leclercq, Tristan and {de Langre}, Emmanuel},
 	month = jul,
 	year = {2018},
 	pages = {77--93},
@@ -293,8 +293,37 @@
 	doi = {10.1016/j.jfluidstructs.2015.10.007},
 	urldate = {2015-12-13},
 	journal = {Journal of Fluids and Structures},
-	author = {Leclercq, Tristan and de Langre, Emmanuel},
+	author = {Leclercq, Tristan and {de Langre}, Emmanuel},
 	month = jan,
 	year = {2016},
 	pages = {114--129},
+}
+
+@article{buchak_clapping_2010,
+	title = {The {Clapping} {Book}: {Wind}-{Driven} {Oscillations} in a {Stack} of {Elastic} {Sheets}},
+	volume = {105},
+	shorttitle = {The {Clapping} {Book}},
+	url = {https://link.aps.org/doi/10.1103/PhysRevLett.105.194301},
+	doi = {10.1103/PhysRevLett.105.194301},
+	number = {19},
+	urldate = {2019-08-18},
+	journal = {Physical Review Letters},
+	author = {Buchak, P. and Eloy, C. and Reis, P. M.},
+	month = nov,
+	year = {2010},
+	pages = {194301},
+}
+
+@article{goldberg_mechanics-based_2020,
+	title = {Mechanics-based model for the cooking-induced deformation of spaghetti},
+	volume = {101},
+	url = {https://link.aps.org/doi/10.1103/PhysRevE.101.013001},
+	doi = {10.1103/PhysRevE.101.013001},
+	number = {1},
+	urldate = {2020-01-05},
+	journal = {Physical Review E},
+	author = {Goldberg, Nathaniel N. and O'Reilly, Oliver M.},
+	month = jan,
+	year = {2020},
+	pages = {013001},
 }
