+++ paper/paper.md
@@ -27,11 +27,11 @@ bibliography: paper.bib
 
 # Summary
 
-Elastic rods are long and thin structures, and are the object of many dynamical systems in nature, from bending and swaying plants [@gosselin_mechanics_2019] to swimming flagellated and ciliated bacteria [@brennen_fluid_1977; @lauga_bacterial_2016] and propulsing carangiform fish [@cheng_continuous_1998]. In industry as well, elastic rods serve as models to study drilling risers in offshore sites and tall chimney stacks in constructions that may vibrate in presence of water or air flows [@paidoussis_fluid-structure_2010]. With the prevalence of rods in mechanical problems, it is important to have a simple to use efficient numerical tool reproducing faithfully the observed dynamics. RodiCS is a Python code that fulfills this task, and simulates the motion of elastic rods under defined forces in general, and their interaction with fluid flows more specifically. While numerous researchers [@violette_computation_2007; @lei_wave_2019; @gosselin_drag_2010; @hassani_large_2016; @leclercq_reconfiguration_2018; @leclercq_vortex-induced_2018; @leclercq_drag_2016; @luhar_wave-induced_2016] spend valuable time in writing *ad hoc* codes with limited verification and validation, RodiCS comes as a versatile and practical open source library that scientific projects can benefit from and build upon.
+Elastic rods are long and thin structures, and are the object of many dynamical systems in nature, from bending and swaying plants [@gosselin_mechanics_2019] to swimming flagellated and ciliated bacteria [@brennen_fluid_1977; @lauga_bacterial_2016] and propulsing carangiform fish [@cheng_continuous_1998]. In industry as well, elastic rods serve as models to study drilling risers in offshore sites and tall chimney stacks in constructions that may vibrate in presence of water or air flows [@paidoussis_fluid-structure_2010], and even to investigate the cooking process of a strand of spaghetti [@goldberg_mechanics-based_2020]. With the prevalence of rods in mechanical problems, it is important to have a simple to use efficient numerical tool reproducing faithfully the observed dynamics. RodiCS is a Python code that fulfills this task, and simulates the motion of elastic rods under defined forces in general, and their interaction with fluid flows more specifically. While numerous researchers [@violette_computation_2007; @lei_wave_2019; @gosselin_drag_2010; @hassani_large_2016; @leclercq_reconfiguration_2018; @leclercq_vortex-induced_2018; @leclercq_drag_2016; @luhar_wave-induced_2016; @buchak_clapping_2010; @goldberg_mechanics-based_2020] spend valuable time and effort in writing *ad hoc* codes with limited verification and validation, RodiCS comes as a versatile and practical open source library that scientific projects can benefit from and build upon.
  
 # Description
 
[…]
