### 2020-09-02 #2577 @mou3adb
Hi @chennachaos. Thank you for these comments. Here are our responses:

Comment 1.) _In essence, this library is a finite element analysis code for studying the response of a single elastic rod modelled with 2-noded elements. The coupling between the fluid and solid is modelled as an external force and van der Pol equation. It does not even model multiple rods and their interactions. The problem tackled here is too simple and the library is too small to be considered for another publication._

Yes, RodiCS is a FEniCS-based solver that simulates the response of a single Kirchhoff rod using interval finite elements embedded in three-dimensional space. The wake-oscillator model is only one example of the ability of RodiCS to compute coupled dynamics, and other phenomenological or reduced-order models should be easily implemented following the framework we propose. Our code combines geometrical non-linearity, the semi-empirical and non-linear drag model, and the distributed wake-oscillators model, features that are of great interest for diverse applications in fluid-structure interaction and beyond. In addition, owing to its versatility, RodiCS is a verified and validated library that could have saved valuable time for many graduate students, among whom those cited below, that wrote _ad hoc_ codes in their research projects from zero. To strengthen the case for the motivation of the work in the paper, we added the following passage in the section `Summary':

> While numerous researchers (Buchak, Eloy, & Reis, 2010; Goldberg &O’Reilly, 2020; Gosselin, de Langre, & Machado-Almeida, 2010; Hassani, Mureithi, & Gosselin,2016; T. Leclercq & de Langre, 2018; Tristan Leclercq & de Langre, 2018; Leclercq & deLangre, 2016; Lei & Nepf, 2019; Luhar & Nepf, 2016; Violette, de Langre, & Szydlowski,2007) spend valuable time and effort in writing _ad hoc_ codes with limited verification and validation, RodiCS comes as a versatile and practical open source library that scientific projectscan benefit from and build upon.

In all these cited papers, graduate students coded _ad hoc_ software with limited capabilities, benefiting from limited verification and validation. By building upon an open-source code which can be further developed by many researchers, deeper can be achieved.

Concerning the multiple rod interactions, the solution of a many-rod problem or an arborescent structure all coupled mechanically is directly amenable with the code. In terms of coupling via the flow, this is not possible with the phenomenological wake-oscillator model that is implemented and is beyond the scope of the software at the moment, although it is an interesting research topic.

Comment 2.) _The theory behind the library is not available and it is not yet peer-reviewed. It seems to be under review for publication in another journal._

The derivation of equations, numerical formulation, code verification and validation, as well as some simulation cases, are described in chapter 7 of my thesis, which I joined a copy in the repository. As for the paper mentioned, it is actually an example of the code usage in an academic study, as recommended in the review criteria, in ‘The JOSS paper’ section.

Comment 3.) _The library is less than two months old. The library does not follow the standard guidelines in organising the files. There is no distinction between the source code and test cases. All files in the scripts folder contain only 2247 lines of code._

The online repository is less than two months old, yet the real starting time of this project is older. I began to work on the project in my own directory in my personal computer, as we did not think about making a publishable library at that time. We are not aware of publication criteria regarding the repository age.

Also, we are not aware of any file organisation standard prescribed in the JOSS guidelines.

Regarding test functions, we added test files with names starting with the prefix `test_`, followed by the type of simulation, either `_static_` or `_dynamic_`, then the simulation case. Each of these files contains a list of input parameters that the user may tune. To make it even clearer, we included in the repository description a section explaining how to run the tests.

The number of lines of code (LOC) respects the length requirements of JOSS mentioned in the ‘Substantial scholarly efforts’ section. RodiCS builds upon FEniCS to achieve much of the calculation capabilities. It makes efficient uses of pre-established packages and codes.

Comment 4.) _The library lacks dedicated documentation describing the various aspects of the code. The thesis linked to the library is not accessible; the link is not working. This is not an issue that contributed to my decision._

We acknowledge this remark, and understand that this was not an issue for the decision following this review. As I mentioned in the response to Comment 2.), I joined a copy of my MSc thesis to the online repository. The theory supporting the code is dedicated in chapter 7.

Comment 5.) _The authors' statement "The traditional method to simulate the motion of elastic rods in finite elements is to solve the elasticity equation in three-dimensions" is not correct. There are well-established rod and beam elements for modelling such a class of problems. Adding external forces to such models in any reasonably established FE software is not difficult._

This is a good point. We replaced "in finite elements" by "in FEniCS" to precise that it is the traditional method used in this platform. To the extend we looked into, we have not found any FEniCS code simulating a transient problem using one-dimensional elements. Broadly, open source finite element solvers rarely use 1D Kirchhoff rod elements embedded in a 3D space. Upon this latter affirmation, we emphasised the importance of RodiCS by adding the following sentence in the first paragraph of the section `Statement of Need':

> Therefore, RodiCS fills a need in open source finite element codes for a way to solve the one-dimensional Kirchhoff rod equation embedded in a three-dimensional space.

We would like to add that it is the difficulty in some commercial software to implement customised loads, especially non-linear, space and time dependent loads, that pushed us to code our own library. For months, we tried different approaches with ANSYS and Abaqus, went over the provided functionalities, but have not found a proper way to include the semi-empirical non-linear drag model for a beam under a constant flow speed, a very basic FSI problem, let alone the coupled wake-oscillator model. With RodiCS, this difficulty is wiped out, and even non-programmer users can simulate time and space dependent, coupled rod dynamics fluently and straightforwardly.

Comment 6.) _The authors' argument for the maintenance of the code, or the lack of it, is weak, considering especially that the framework is based on the 2-noded finite element. I do not see any motivation for using FEniCS. I think that the use of FEniCS for this problem is overkill. Why not implement 1-d basis functions in the code itself?_

The argument is one of _ad hoc_ coding versus one made to be part of a bigger environment. Open source codes that can achieve some of the capabilities in RodiCS do exist. For example, one such code we tried functioned only on a precise distribution of Linux, making use of 5 version specific libraries, which all had to be properly linked. The code, which was less than 5 years old, had instructions explicitly mentioning that it was not maintained and that assistance was not available for installing or running it. In this sense, we believe that a code based on Python and FEniCS, an open source FEM with an active community, is an important advantage for RodiCS. The use of FEniCS imposes a standard in the definition of meshes, making possible mesh generation with external meshing software, or combination of the elements in RodiCS with other types of finite elements. This modular approach is definitely a feature. We stressed this point in the paper by replacing the second part of the second paragraph of `Statement of Need' with the following:

> These drawbacks were incentives for the authors of the present paper to think of writing their own code based on the FEniCS project, an open source finite element solver with an active community, continuously developed, and cross-platform; all significant advantages. Also as an essential modular feature, FEniCS imposes a standard in the mesh definition, making possible either the mesh generation with external meshing software, or the combination of elements in RodiCS with other types of finite elements. Finally, written in Python, RodiCS is inclusive and accessible for users of all programming levels.

Comment 7.) _There is only one time integration scheme, the backward Euler scheme, and it is only first-order. There are numerous well-established second-order time schemes._

At this point, first order time integration worked for the purpose we had and the precision we desired. Implementing other time integration schemes is on the feature development list. We clarified more this proposition in the ‘Contribution’ section of our repository description.

Finally, I would like to mention some minor changes we added in the paper:

- 'and even to investigate the cooking process of spaghetti' of the section 'Summary',
- 'non-linear' before 'drag formulation' in first paragraph of section 'Description',
- 'in chapter 7 of Boudina (2020)' in the last sentence of section 'Description',
- 'verification' in the last line of the first paragraph of section 'Description',
- 'cross-platform' in the last line of second paragraph of section 'Statement of Need',
- 'semi-empirical non-linear' in the last line of section 'Complex Dynamics',
- the financial support 'RGPIN-2019-05335' in the acknowledgements.
