### 2020-09-07 #2577 @chennachaos
Hi @mou3adb. Thank you for your response! Your MSc thesis is great; however, your arguments are not convincing. I do not recommend this library in its current form for publication in JOSS.

The primary issue in recommending this library for publication in JOSS is the lack of content. There is not much one can do with this library. The type of problems that can be readily solved using this library is very limited. In fact, it is limited to a problem with a single rod with prescribed force and coupled to another discrete oscillator model. This code is not for direct numerical simulations of the interaction of thin rods with fluid flow. Considering the dimensionality of the problem, *this, in its current form, is not such a challenging problem that a library dedicated to it deserves a publication in JOSS*.

A thorough look at your thesis, and the concerned references therein, regarding the problem of vortex-induced vibrations, reveals that there is no quantitative comparison of numerical results for the dynamic case. How accurate is the lift-force model for simulating complex models such as non-uniform flows, multi-directional flows, geometries with multiple bifurcations such as the one shown in Figure 3.2. in your thesis? These kinds of questions concerning the theory should be answered before considering the library for publication. Extension of this framework to other complex problems requires a significant effort, in which case, my question is, why should the authors of this library get the credit when others need to add all the features/models they want to simulate? Extensions should only constitute a fraction of the work required to develop the library.

I sympathise with you regarding the lack of maintenance of research codes. However, your arguments on the use of FEniCS are not particularly convincing. I see no motivation for basing this code on FEniCS or any other FE core library for that matter.
By relying on FEniCS, you are imposing unnecessary requirements on the user to install that library, considering especially that all that is needed for this problem is the basic 2-noded finite elements which can be hand-coded into the library itself. I don't see what other features of FEniCS you can use for the extension of this particular code. However, the book-keeping and maintenance aspects of the library are secondary here.

Last but not the least, have a look at other libraries published in JOSS to get an idea of (i) the structure of source code, (ii) documentation and (iii) examples and test cases.
