### 2020-09-08 #2577 @diehlpk
`There is not much one can do with this library. The type of problems that can be readily solved using this library is very limited. In fact, it is limited to a problem with a single rod with prescribed force and coupled to another discrete oscillator model. This code is not for direct numerical simulations of the interaction of thin rods with fluid flow. Considering the dimensionality of the problem, this, in its current form, is not such a challenging problem that a library dedicated to it deserves a publication in JOSS.`

Hey, @mou3adb. I think this quote from @chennachaos is my major concern. To back up my decision, I like to quote JOSS's Scope & submission requirements

```

JOSS submissions must:

1.  Be open source (i.e., have an OSI-approved license).
1. Have an obvious research application.
1.  Be feature-complete (no half-baked solutions) and be designed for maintainable extension (not one-off modifications).
1. Minor 'utility' packages, including 'thin' API clients, and single-function packages are not acceptable.
```
Your submission clearly addressed the first two points. However, I do not see how the other two points are addressed.

`1. Minor 'utility' packages, including 'thin' API clients, and single-function packages are not acceptable.`

I think your package in the current state is a minor 'utility' packages and it is more a single-function, meaning that one can only solve one problem (single rod) and can just change a limited set of parameters. I do not think that one could use your code in the current state without major modifications to solve any other problem than your single rod example problem. Which relates to the third point in the list above. I do not think that your code in the current state is feature-complete and it is more a showcase/example for your research. However, for JOSS we like to see code, which can be easily used by others to solve research problems. Again, I agree with @chennachaos one would have to do major modification to the code to use it for direct `numerical simulations of the interaction of thin rods with fluid flow`.

I would like to reject this paper in its current form due to a minor 'utility' package. However, I like to encourage you make the code feature-complete and add more functionality, e.g. make the simulation more configurable and add multi rod support. We look forward to your future submissions to JOSS.
