+++ paper/paper.bib
@@ -327,3 +327,19 @@
 	year = {2020},
 	pages = {013001},
 }
+
+@article{mukundan_monitoring_2009,
+	series = {Bluff {Body} {Wakes} and {Vortex}-{Induced} {Vibrations} ({BBVIV}-5)},
+	title = {Monitoring {VIV} fatigue damage on marine risers},
+	volume = {25},
+	issn = {0889-9746},
+	url = {http://www.sciencedirect.com/science/article/pii/S0889974609000383},
+	doi = {10.1016/j.jfluidstructs.2009.03.003},
+	number = {4},
+	urldate = {2020-09-08},
+	journal = {Journal of Fluids and Structures},
+	author = {Mukundan, H. and Modarres-Sadeghi, Y. and Dahl, J. M. and Hover, F. S. and Triantafyllou, M. S.},
+	month = may,
+	year = {2009},
+	pages = {617--628},
+}
