### 2020-09-09 #2577 @mou3adb
Hi @diehlpk and @chennachaos. Thank you for these replies. Please consider the responses below, we clarified the points that we think raised confusion.

For the claim that `There is not much one can do with this library`, we say first that RodiCS is not a 'single-function package' nor a 'thin API client.' RodiCS can reproduce the modelling results of the following papers (Buchak, Eloy, & Reis, 2010; Goldberg & O’Reilly,2020; Gosselin, de Langre, & Machado-Almeida, 2010; Hassani, Mureithi, & Gosselin, 2016;T. Leclercq & de Langre, 2018; Tristan Leclercq & de Langre, 2018; Leclercq & de Langre,2016; Lei & Nepf, 2019; Luhar & Nepf, 2016; Mukundan, Modarres-Sadeghi, Dahl, Hover, &Triantafyllou, 2009; Violette, de Langre, & Szydlowski, 2007). This list is far from exhaustive. These papers deal with geometrically non-linear static deformations of structures under flow, dynamics of slender structures subject to wave action, and vortex-induced vibrations. RodiCS has not limited applications. It not only goes beyond reproducing each of these effects, it allows combining all of them together. Also in the 'Summary' section we pointed the diversity of problems and exciting topics which RodiCS should contribute to. Ignoring the importance of the single rod problem means ignoring a whole legacy in FSI; it has many relevant applications in nature and industry alike (pipes, risers, antennas, terrestrial and aquatic vegetation, microbial locomotion, hair models, whiskers, to name but a few). Therefore we cannot really claim that RodiCS 'has minor utility.' We also said in the previous response that the many-rod problem is amenable with the code.

Regarding the following quote `This code is not for direct numerical simulations of the interaction of thin rods with fluid flow`, all the purpose of RodiCS is to avoid direct numerical simulations. We do not see any requirement in JOSS to include DNS in mechanical codes to be published.

`Considering the dimensionality of the problem, this, in its current form, is not such a challenging problem that a library dedicated to it deserves a publication in JOSS.`
Writing an open source code for 1D Kirchhoff rods embedded in 3D space is challenging in FEniCS. No such example is found in the web. This is an essential argument why we want to publish the paper; we do not want another endless list of authors spend months hand-coding _ad hoc_ scripts with rudimentary and basic functions. Instead, researchers and graduate students can use RodiCS and save more time. The guidelines of the JOSS state clearly that any paper `making addressing research better: simpler, easier, faster` is a substantial scholarly effort recommended for publishing. Moreover, always according to the guidelines, RodiCS `has already been cited` in an academic study which we mentioned in the section 'Complex Dynamics', and we expect it `likely to be cited by other researchers` in the future.

`There is no quantitative comparison of numerical results for the dynamic case. [...] These kinds of questions concerning the theory should be answered before considering the library for publication.`
We did not develop any theory here. We implemented different theories in a single code. We do not believe it is necessary to validate each of these theories. In terms of quantitative numerical results, RodiCS successfully reproduces the dynamics of a wave-swept thin plate observed in the experiments of Leclercq and de Langre (2018), as shown in Figure 7.6 of my thesis. The paper of the wake-oscillator model earns 600 citations, and cited several times every single year since its publication, which is impressive, and absolutely not 'of minor utility.' We find it disappointing not to have any open source code implementing such an important model applied in series of experimental and numerical studies. Any doubts on or improvements of the model itself are worth discussing with the corresponding author of that paper.

`Extension of this framework to other complex problems requires a significant effort, in which case, my question is, why should the authors of this library get the credit when others need to add all the features/models they want to simulate? Extensions should only constitute a fraction of the work required to develop the library.`
Extension to other features does not require a significant effort. We did the significant effort already. Again, we do not want future researchers and graduate students to repeat this significant effort. They only need to enter the force expressions as their figure in their drafts. We already mentioned this advantage in the 'Description' section.

`I see no motivation for basing this code on FEniCS or any other FE core library for that matter. By relying on FEniCS, you are imposing unnecessary requirements on the user to install that library, considering especially that all that is needed for this problem is the basic 2-noded finite elements which can be hand-coded into the library itself. I don't see what other features of FEniCS you can use for the extension of this particular code.`
Hand-coding existing modules is time-killing and counterproductive; building upon them is enriching. Here we simply use the same justification as Bergersen, Slyngstad, Gjertsen, Souche, and Valen-Sendstad (2020) who published their code in JOSS recently: basing the code on FEniCS allowed us to write it _"in only a [couple of hundred] lines of high-level Python code, in contrast to tens of thousands of lines of low-level C++ code. This provides full transparency and a unique opportunity for researchers and educators to modify and experiment with the code, while still providing out of the box entry-level high-performance computing capabilities. Furthermore, because of the close resemblance between mathematics and code in FEniCS, users can make additions or modifications with ease."_ This latter phrase was put under light in our paper by indicating explicitly the Unified Form Language (UFL) as a simplifying and powerful asset of FEniCS. With UFL, the range of mechanical problems RodiCS can tackle is easily extended. Finally, installing FEniCS is straightforward, fortunately we do not impose anything on users.

`However, the book-keeping and maintenance aspects of the library are secondary here.`
Not sure about that. Maintenance is a vital component commercial software and open source codes rely on. The DER code is an interesting code, but is now let down and forgotten just for being unmaintained, which is a pity. In contrast, FEniCS has a large and active community, and users can report any relevant issue in the forum and expect an answer.

A minor change to the paper: replaced 'In FEniCS, such a code...' instead of 'Such as code...' in the Statement of Need section.
