### 2020-09-09 #2577 @ilyasst
I am sorry about the delay (classes 😅). I have filled my review section.
Thank you for the previous review and responses @chennachaos and @mou3adb. I believe that in this current state, this code does not seem ready for publication in JOSS. The main comment I have is regarding the current way the code is written (structure and format). The code presented here is a collection of scripts that contain functions that cannot really be re-used to solve different kinds of problems, they can only be used exactly "the way they are written". I do not think that in this state this code fits the definition of a library, it could however be a sound basis for future work that would refactor it and extend it to make a library (and package) out of it. Please understand that my comments will mainly address the code. Although I think that your work is a relevant contribution (thesis and publication), that is however independent from the evaluation of the python code written to support that work.

I am not certain that this code fits the JOSS guidelines in its current format. The repository presented here contains several folders among which two are of interest, `scripts` and `xml_files`. The scripts folder contains a collection of scripts. The scripts contain several hard-coded values that do not have comments or documentation that explains their role (please see my comment about docstrings in the following section).
Are the files that start with `test` actually examples ? I do not believe that those are "tests" (I was expecting something like this https://docs.python.org/3/library/unittest.html), however they look like examples.

As a user, I have no idea how to solve my own problem, meaning how to re-use this collection of scripts to solve one of my own problems. It should be possible to easily import and re-use a library to solve a user-specific problem. I have spent quite some time looking through the proposed code and documentation and have no idea how to do so. Should I use the tests to learn how to do so ?

This code feels like an ongoing work that still requires some refactoring and improvement. The functions sometimes feels like wrappers around scripts that are used as compartments or sections in a document more than ways to reduce repetition and make the code re-usable. The code still contains "comment-separations" :
```
#####################################
#===============================
```
I also notice something I could not really understand about docstrings: for example, in the `dynamic.py` file there is a quite good docstring for the run_dynamic function; howeve, in the  `static.py` file there is no docstring at all. By comparing the files `dynamic.py` and `static.py` it appears that the two files contain several exact same sections, they even contain the exact same fenics functions (such as root or tip). Although functions are used in this code, all these lines are repeated. I believe that the author could use functions here, and in several other places, to avoid repetition. In addition, those function could be used by several modules.
The structure of the code also feels like a work in progress, the whole code should be reviewed while trying to consider more problems, make it easy for the user to provide inputs, and focus on avoiding repetitions and building "more general" functions. The code should not be presented as a loose collection of scripts.

### Points that I would recommend the authors should address:

- This collection of scripts cannot be "installed". It can only be copy-pasted. The review checkpoints mention installation several times. I recommend that this code should be re-written to be presented, in its format at least, as an installable package/library (please see https://stackoverflow.com/questions/15746675/how-to-write-a-python-module-package)

- I would strongly recommend that future users and potential contributors should easily have access, in a concise and clear format, to all the necessary documentation to understand the work presented here. I do not believe that they should have to go through your thesis to do so. The documentation is presently quite thin and requires additional work.

- Consider using `docstrings` to comment and document your code. It is already quite hard to understand how to use collections of scripts, it is certainly harder when there are no comments guiding them. At this stage, I cannot say that this code has a documentation. I could not find it, and the README is definitely not enough. The comments are scarce and are not clear enough.

- A better `README.md` should be provided. The last title section of the Usage section is "Visualisation" and its content is "The visualisation and graphical tools are written in the file `visualisation.py`." I think that the author will agree with the reviewer that this is not very helpful. Now if we open the `visualisation.py`, the comments are very scarce and the functions are extremely specific (and rigid) to this problem. The function `prepare_frame(ax, lambd=15, phi=50)` is extremely specific to this work. In addition, other files in the scripts folder also contain their own graphics functions to plot things. So what is `visualisation.py` for ? This is confusing and makes it hard for the user to understand how the code works and what it is it actually does when.

- The section "Usage" in `README.md` does not actually explain how to use the code. I think that is is mostly because, in its current state, this repository was not provided to be used, but mostly to be executed. The "Usage" section should not only tell the user how to execute the code as it is, it should explain to the user how they can solve their own problem.

- I do not think that the functions provided in this code have proper tests in place. Could you please double check that ?

- An automated package management solution should be implemented, please consider including a `requirements.txt` file.

- Consider using logging instead of prints although this is optional (https://docs.python.org/3/howto/logging.html)

The last point I would like to add is that, if I actually understood how this collection of scripts actually works, meaning that the user should write a file similar to one of the test files, then this collection of scripts should make it easy for the user to write such a script. Currently I do not think that it is the case.

Finally, in the paper the following point is put forward to justify this work `Because the DER code is written in C++ and has an intricate file structure, only programmer users are able to modify the source code, excluding thus users new in coding or having basic knowledge in programming.` Although I agree with the general idea that python might have a "nicer" syntax than C++, that does not mean that re-writing a code in python automatically makes it more accessible. That is only true if the code is actually written following several guidelines (the python standard documentation, pep-8, ...). Currently, I do not think that this code is accessible, it is however a good first step to build such a library.

_Disclosure: I was recently a postdoc in the same lab as @mou3adb (although not in the same field) and I have collaborated with his supervisor on a publication in the last two years. However, I was not involved in this work at all. I have discussed this with @diehlpk._
