### 2020-09-09 #2577 @diehlpk
@mou3adb having the second review, I really think that the code is not in the shape to be considered for a publication in JOSS. Note that a JOSS publication is about the source code, in your case the Python code. I do believe that the research you have done is good and interesting, however, you got credit for that in your master thesis and the publication under review. I recommend to not accept the code in its current form. However, I do think, if you wait for some time, improve the code structure, add documentation, do not have hard-coded values, and make it more clear how users should use your code, I think you should resubmit to JOSS.
