+++ paper/paper.bib
@@ -1,345 +1,0 @@
-@book{audoly_elasticity_2010,
-	address = {Oxford ; New York},
-	title = {Elasticity and {Geometry}: {From} hair curls to the nonlinear response of shells},
-	isbn = {978-0-19-850625-6},
-	language = {English},
-	publisher = {Oxford University Press},
-	author = {Audoly, Basile and Pomeau, Yves},
-	month = aug,
-	year = {2010}
-}
-
-@book{landau_theory_1986,
-	title = {Theory of {Elasticity}},
-	language = {en},
-	publisher = {Pergamon Press},
-	author = {Landau, L. D. and Lifshitz, E. M.},
-	month = jan,
-	year = {1970},
-	keywords = {Science / Mechanics / General, Science / Physics / General}
-}
-
-@article{alnaes_fenics_2015,
-	title = {The {FEniCS} {Project} {Version} 1.5},
-	volume = {3},
-	copyright = {Copyright (c) 2015 Archive of Numerical Software},
-	issn = {2197-8263},
-	url = {https://journals.ub.uni-heidelberg.de/index.php/ans/article/view/20553},
-	doi = {10.11588/ans.2015.100.20553},
-	language = {en},
-	number = {100},
-	urldate = {2020-03-26},
-	journal = {Archive of Numerical Software},
-	author = {Alnæs, Martin and Blechta, Jan and Hake, Johan and Johansson, August and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and Ring, Johannes and Rognes, Marie E. and Wells, Garth N.},
-	month = dec,
-	year = {2015},
-	note = {Number: 100},
-}
-
-@article{alnaes_unified_2014,
-	title = {Unified form language: {A} domain-specific language for weak formulations of partial differential equations},
-	volume = {40},
-	issn = {0098-3500},
-	shorttitle = {Unified form language},
-	url = {https://doi.org/10.1145/2566630},
-	doi = {10.1145/2566630},
-	number = {2},
-	urldate = {2020-06-27},
-	journal = {ACM Transactions on Mathematical Software},
-	author = {Alnæs, Martin S. and Logg, Anders and Ølgaard, Kristian B. and Rognes, Marie E. and Wells, Garth N.},
-	month = mar,
-	year = {2014},
-	pages = {9:1--9:37},
-}
-
-@article{facchinetti_coupling_2004,
-	title = {Coupling of structure and wake oscillators in vortex-induced vibrations},
-	volume = {19},
-	issn = {0889-9746},
-	url = {http://www.sciencedirect.com/science/article/pii/S0889974603001853},
-	doi = {10.1016/j.jfluidstructs.2003.12.004},
-	number = {2},
-	urldate = {2019-01-28},
-	journal = {Journal of Fluids and Structures},
-	author = {Facchinetti, M. L. and {de Langre}, Emmanuel and Biolley, F.},
-	month = feb,
-	year = {2004},
-	pages = {123--140},
-}
-
-@article{leclercq_reconfiguration_2018,
-	title = {Reconfiguration of elastic blades in oscillatory flow},
-	volume = {838},
-	issn = {0022-1120, 1469-7645},
-	url = {https://www.cambridge.org/core/journals/journal-of-fluid-mechanics/article/reconfiguration-of-elastic-blades-in-oscillatory-flow/93C943F932866F3DCAFE8C53F34540F7},
-	doi = {10.1017/jfm.2017.910},
-	language = {en},
-	urldate = {2018-01-30},
-	journal = {Journal of Fluid Mechanics},
-	author = {Leclercq, T. and {de Langre}, Emmanuel},
[…]
