+++ README.md
@@ -16,6 +16,8 @@ This code was written during my Masters research project at Polytechnique Montr√
 
 Rod simulations are the subject of Chapter 7. You'll find the Kirchhoff equations, the mechanical and fluid-dynamical load expressions, verification and validation of the code, and finally simulation cases. I also included in it an appendix on the FEniCS implementation in Python, describing the syntax and functions used in the code.
 
+Please do check the document `RodiCS_a_finite_element_solver_of_Kirchhoff_rods_under_fluid_flow_and_more.pdf`, it summarises the global perspective of the code and lists some of its advantages.
+
 ## Dependencies
 
 RodiCS works on Python (version 3.8.2), and requires FEniCS (version 2019.1.0), as well as NumPy (version 1.17.4) and Matplotlib (version 3.1.2) modules. Only for data postprocessing or code validation, you'll also need SciPy (version 1.4.1).
@@ -57,22 +59,22 @@ The function `run_dynamic` requires the same argument as in the static case, in 
 #### Wake-oscillator model
 The function `run_wake_oscillator` in `wake_oscillator.py` is written in the same fashion as `run_dyamic`. Because the wake-oscillator model assumes that the vortex-induced lift is an additional unknown governed by the van der Pol equation (see [Facchinetti et al. (2004)](https://www.sciencedirect.com/science/article/abs/pii/S0889974603001853)), we consider a new mixed space solution with additional subspaces, and the van der Pol equation is included in the final variational form.
 
-### Tests
-In each test case, you can open the file and tune the parameters as you wish. You can either launch the following command in terminal
+### Examples
+In each example, you can open the file and tune the parameters as you wish. You can either launch the following command in terminal
 ```
-python3 test_name.py
+python3 example_name.py
 ```
 or execute it in a Python interactive editor such as [Spyder](https://www.spyder-ide.org/).
 
 #### Static case
-The file `test_static_distributed.py` gives the deformation of a rod under a uniformly distributed load. After the simulation has done, a window will open with a 3D view of the rod.
+The file `example_static_distributed.py` gives the deformation of a rod under a uniformly distributed load. After the simulation has done, a window will open with a 3D view of the rod.
 
-Another test is in `test_static_drag.py`, which simulates a rod under static drag for different flow speeds. A figure will appear showing the deformations with a top view.
+Another example is in `example_static_drag.py`, which simulates a rod under static drag for different flow speeds. A figure will appear showing the deformations with a top view.
 
 #### Transient case
-We suggest three tests. The files `test_dynamic_ramp_up.py` and `test_dynamic_oscillatory.py` simulate the motion of a rod under a ramp up flow speed and an oscillatory flow. The file `test_wake_oscillator_ramp_up.py` simulates the motion of a rod under a ramp up flow speed, arising vortex-induced vibrations.
+We suggest three examples. The files `example_dynamic_ramp_up.py` and `example_dynamic_oscillatory.py` simulate the motion of a rod under a ramp up flow speed and an oscillatory flow. The file `example_wake_oscillator_ramp_up.py` simulates the motion of a rod under a ramp up flow speed, arising vortex-induced vibrations.
 
[…]
