### 2021-05-04 #2577 @diehlpk
> Dear Patrick, Yes you can close the review. Mouad's paper was published in the Journal of Fluid Mechanics. The paper cite a zenodo DOI for the code https://zenodo.org/record/4023287#.YIxHb7VKiUk Hopefully, I get the manpower in the future to build on this code and perhaps submit it again to JOSS in the future. Thanks for your time, Frédérick Gosselin Professeur Agrégé, Département de Génie Mécanique, Polytechnique Montréal, Canada Laboratoire de Mécanique Multi-Échelles (LM2) http://www.fgosselin.com Cellulaire : (514) 713-6245 Bureau : (514) 340-4711 ext. 3747
> […](#)
> On Fri, 30 Apr 2021 at 12:24, Patrick Diehl ***@***.***> wrote: @mou3adb <https://github.com/mou3adb> I will close the review next week, since I have not heard back from you since January. — You are receiving this because you were mentioned. Reply to this email directly, view it on GitHub <[#2577 (comment)](https://github.com/openjournals/joss-reviews/issues/2577#issuecomment-830209290)>, or unsubscribe <https://github.com/notifications/unsubscribe-auth/AKYVZYCEFILZCGNGP5ZS4T3TLLKSDANCNFSM4QARJNVA> .

Sure, thing. I look forward to the updated code base. Please feel free to suggest me as the editor for any future submisison.
